job_alerts/requirements.txt
django 
djangorestframework
django-cors-headers
requests
beautifulsoup4
dateparser
djangorestframework-simplejwt
